Bundle support for RHEL, SUSE platforms
This is the beginning of bundle support (for now, just RHEL and SUSE, although as you look this over, you'll see hooks for AIX, HPUX, and Sun as well).  This will work for ULINUX as well, although in the end ULINUX will have it's own file due to different requirements (multiple kits based on SSL version).  But that will come later.

Along for the ride is a trivial change to metaprovider.cpp: Print the locale.  I caught this (while working on an RFC) in the logs, and realized OM with OMI won't show this any longer.  So adding it back in (just for metaprovider - that seems sufficient).