Bundle support for Solaris platform
I'm not thrilled with the Solaris-specific tests in the primary.skel file.  While I'm not thrilled, I don't think it'll make maintenance much harder.  However, if folks disagree, comment!  I can do a solaris.skel file in addition to a primary.skel file, and use solaris.skel for Solaris platforms.

(Code reviewers agreed that it doesn't make maintenance much harder, so including the changes as is)