Fix HandleConfigFiles to be under 'set +e' context on Linux and SunOS.
This was a bug fix in 2012R2_URNext for CU3 of OM, now being applied to the working branch.

Kind of a pain to insure with the new installer (what the working branch uses).  Needed to do a build everywhere (pbuild), then look at resulting PostInstall files for each of the platforms.  Only a problem on Linux and SunOS.