[CDM] Ignore log file events during maintenance mode (and add other MOF changes for WI 634777)
These changes backs out the prior mechanism (initialize flag when reading log files), and have added one new Invoke method:

ResetStateFile() will reset a particular logging state file for a specific logfile

While there isn't an invoke method for this, there is code in scxlogfilereader to essentially perform ResetAllStateFiles() to reset all logging state files for all monitored logfiles under an account. This is called directly by the MP on Windows.

This includes the latest design changes as discussed with Kris Bash.  Namely:

1. Two ways to reset log file state: on demand (resetOnRead = false), and on the next time logfiles are read (resetOnRead = true). When in maintenance mode, logfiles are never read with GetMatchedRows,
2. Removed provider to ResetAllStateFiles, as this will never be invoked by provider. Instead, scxlogfilereader will be invoked directly to the MP. When invoking scxlogfilereader to reset all state files, we require a parameter to state if ResetOnRead should be set. This is specified as 1/true (to ResetOnRead), or 0/false (to not ResetOnRead).

To make this work, I needed to redesign the parser in scxlogfilereader to be a lot more intelligent about what was run, and when.