Resolve memory leak in RunAs provider (SCXThread object leaking)
Commit hash 0921ef7 (back in 2013) incorporated a memory leak of the
SCXThread object. This was "fixed" in f004d08, which was supposed to
roll back other changes to the SCX_OperatingSystem_Class_Provider.cxx
and then fix the bug.

However, it turns out that the fixes never made it past the rollback.
This commit fixes those leaks for real.