Remove and reinstall packages if upgrade fails
Previously, we used a "install all packages as a list" (commit hash 6a0acc1)
to shake an upgrade bug from kit -240 to later. But this didn't pan out
consistently. So now if upgrade fails for some reason then remove underlying
packages and reinstall them (forcing removal in case of dependencies).
This seems to work reliably on both rpm and dpkg systems.