Report omsagent version, if installed, in SCX_Agent's kitVersionString
SCOM CB release will ship omsagent bundle instead of scx bundle on
Linux platform. SCOM console needs to report omsagent version as
the installed agent version.
kitVersionString's value in SCX_Agent enumeration is shown in SCOM
console. Change is made to check if omsagent install info file is
present or not. If present it will be parsed to get the value of
install version and install time. If not scx install info file
will continue to be used.

This change is only for SCOM CB release. 2016 UR's will remain
on current implementation.