Open port 1270 on install if it was open at uninstall
While upgrading omsagent, azure linux extension calls --remove
followed by --install, instead of directly calling --upgrade.
This sequence wascausing port 1270 to close as we close the
port on uninstall and do not open it by default on install.

To support this scenario we will now check if port 1270 is
open at uninstall. If yes we will create a marker file in
scx conf directory. On install if the marker file is present
then port 1270 will be opened.

As marker file is created in scx conf directory, it will get
deleted in --purge call.