Update omikey ownership to omi:omi during scx upgrade.
When upgrading from 2012 R2 to latest scx omikey ownership
is not getting updated to omi:omi. Due to this omi is not
able to read the key file, hence port 1270 is getting closed.
To address this scxsslconfig is updated to check and update
omikey ownership if key file alreday exists.